fixed bad merge conflicts

+++ Node-Beaver/Node-Beaver.cydsn/can_manager.h
@@ -21,14 +21,14 @@ void can_init();
 
 uint8_t can_test_send();
 
-void can_get(DataPacket* data_queue, uint8_t* data_head, uint8_t* data_tail);
+void can_get(DataPacket* data_queue, uint16_t* data_head, uint16_t* data_tail);
 
-void wrap_data_queue(uint8_t* data_head, uint8_t* data_tail);
+void wrap_data_queue(uint16_t* data_head, uint16_t* data_tail);
 
-inline void process_default(DataPacket* data_queue, uint8_t* data_head,
-	uint8_t* data_tail);
+inline void process_default(DataPacket* data_queue, uint16_t* data_head,
+	uint16_t* data_tail);
 
-inline void process_throttle(DataPacket* data_queue, uint8_t* data_head,
-	uint8_t* data_tail);
+inline void process_throttle(DataPacket* data_queue, uint16_t* data_head,
+	uint16_t* data_tail);
 
 #endif
